fix mouse interaction and dragging while LabelWithTooltip is truncated
Summary:
LabelWithTooltip loads a mouse area when the label is truncated, which interferes with other mouse areas.
Set accepted buttons to Qt.NoButton so that it does not interfere anymore

Test Plan:
Resize application window such that the title labels in the playlist get truncated.
Use mouse to click or drag the entry while hovering the title label.

Reviewers: #elisa, mgallien

Reviewed By: #elisa, mgallien

Subscribers: mgallien

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D8729