reset stackview to initial item when selected via ViewSelector
Summary: when reselecting e.g. the all albums view with the view selector, the stackview shows a previously openend album. Reset to the initial item when the view selector is used

Test Plan: go to all albums and open an album. then select the all artist view and then select the album view again

Reviewers: #elisa, mgallien

Reviewed By: #elisa, mgallien

Subscribers: mgallien

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D8876