enhance content view navigation
Summary:
The backspace key is now also accepted for back navigation in e.g. the artist album view.
A mouse area is added to MediaBrowser to allow back navigation with the mouse backbutton.
Focus is forced on the stackview when arist or album is opened. Without this, the back key navigation is only working when it is actively focused by the user, e.g. by a mouse click

Reviewers: #elisa, mgallien

Reviewed By: #elisa, mgallien

Subscribers: mgallien

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D8895