implement a metadata view for tracks
Summary:
implements a metadata view for the albumview and alltracksview. metadataview is based on a popup and currently read only.
can be accessed by the new info button for each track. the view is ready for more metadata than currently is available, but does autoresizing depending on what is available
Fixes T6345

Reviewers: #elisa, mgallien

Reviewed By: #elisa, mgallien

Subscribers: aacid, ltoscano, kde-i18n-doc

Tags: #elisa, #vdg

Maniphest Tasks: T6345

Differential Revision: https://phabricator.kde.org/D9696