Change color of elapsed time in seek/volume bars
Summary:
Use a neutral color for the elapsed time in the Media Player control's bars.
This follows the original mockup (https://community.kde.org/File:Banagarang_Music_Mockup1a.png)

I think it also looks better because it limits the colors in the header area to the cover's.

{F5662241}

Reviewers: #elisa, mgallien

Reviewed By: #elisa, mgallien

Subscribers: mgallien

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D9964