Aggressively look for cover files
Summary:
If you are like me and accumulated music over years, the conventions for
the file names of those changed over time, we need something more
aggressive.

So now it covers (hehe) files containing cover, folder and front in
their name (with upper case initial or not), it also looks for files
containing the album name in their name (with and without spaces).

It still doesn't catch all my covers, but I think we'll need to refactor
that piece of code to do better. That's for another attempt when I get
the time to do a proper results sorting. :-)

Reviewers: mgallien, ngraham

Differential Revision: https://phabricator.kde.org/D12472