use rating and comment stored in metadata tags
Summary:
if no xattr rating and comment are found, use the
data stored in the metadata. First step for T5375
CCBUG: 395735

Test Plan:
Create an audio file with rating written into the metadata.
The rating shows up in Elisa.
If xattr rating is set through dolphin, this rating is shown instead.
Same for comment

Reviewers: mgallien

Reviewed By: mgallien

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D16362