Global shortcut for play/pause action using spacebar key
Summary:
Assign SpaceBar key as the global shortcut for play/pause action.  Most of the media player out there already use spacebar as the global shortcut for play/pause action.  So,  it would be good to do the same for Elisa.
Modified ElisaMainWindow.qml to add the shortcut component.
Modified MediaPlayerControl.qml to prevent elements from getting focus as it causes interference with the shortcut keypressed events.

BUG:    392989
FIXED-IN:    elisa  0.3.80

Reviewers: #elisa, ngraham, mgallien

Reviewed By: #elisa, ngraham, mgallien

Subscribers: shubham, mgallien, ngraham

Tags: #elisa

Differential Revision: https://phabricator.kde.org/D17994