When not displaying rating, leave more room for song title in playlist
Summary:
Currently the Playlist's ratings widget reserves space for itself for unrated songs,
 which causes song titles to get elided even when there's plenty of space.

This patch hides the ratings widget for unrated songs, allowing the song title label
to have more space and not be elides as often.

This could become a problem in the future if there is ever a desire to set ratings
interactively by clicking on the ratings widget, but that's not implemented so for now
there's no issue (and probably a hover-based UI could make sense for that anyway).

BUG: 406480

Test Plan:
Now song titles for unrated songs have plenty of space, and rated songs display the same:

{F6770576}

Reviewers: mgallien, #elisa

Reviewed By: mgallien, #elisa

Differential Revision: https://phabricator.kde.org/D20511