[RFC] Implement KDE-style header toolbars and statusbars for Playlist and Context view
Summary:
This patch is basically a design proposal in patch form.

Right now, each view's header is not visually separated at all from the content
view below it, and all header areas have different visual styles, some of then
being very tall and taking up a lot of space that could be used for more content.
The context view's flickable is adjusted to take advantage of this, resulting in
a greatly increased amount of vertical space when the lyrics portion is long.

This patch/design proposal implements a new style of header and statusbar
that looks more "KDE-style", for lack of a batter term. It mimics what toolbars
in Kirigami and other QML apps look like. Advantages include:
- Improved visual consistency with modern KDE apps
- Visually pleasing separation from content area
- Scrollable content doesn't appear to get "cut off" under an invisible item anymore
- More compact headers leave more room for content
- Code simplification from using a re-usable component rather than multiple custom views

Test Plan:
Playlist, before: {F6878549}
Playlist, after: {F6878548}

Playlist + context view, before: {F6878551}
Playlist + context view, after: {F6878550}

This patch implements the new header style for only the Playlist and Context View.
If the proposal is approved and the patch lands, I can do the main view too, which
is more complicated to port and will require changes to the autotests as well.

Reviewers: mgallien, #elisa, #vdg

Reviewed By: mgallien, #elisa

Differential Revision: https://phabricator.kde.org/D21676