Make it easier to click on a new position for the playback and volume sliders
Summary:
Because Elisa uses customized visuals for its sliders, various nice features of the
default implementation need to be re-implemented. One example is mouse wheel/scroll
handling, which was done in D22332. Another example is expanded click areas, which is
the target of this patch.

This patch makes the effective click targets fill the entire available space. This makes
it vastly easier to move the slider to a new position with a click.

BUG: 408703

Test Plan:
- It's 1000% easier to click to get a new slider position
- Still looks good with Breeze light and dark color schemes

Reviewers: #elisa, #vdg, mgallien, astippich

Reviewed By: #elisa, mgallien

Differential Revision: https://phabricator.kde.org/D22951