fix import of multi-valued data from existing Baloo db
Summary:
Values like the track artist can ba represented by a QStringList inside a QVariant (probably linked to an old baloo database).
In this case, the value imported in Elisa database is empty (a string list with multiple elements converted to a string gives an empty string).

Test Plan: with this patch, we get a correct string with all the values

Reviewers: #elisa, astippich, ngraham

Reviewed By: #elisa, ngraham

Differential Revision: https://phabricator.kde.org/D22972