allow database to contain and get tracks without metadata
display only the filename

I plan to add on top the work done by astippich that uses a sligthly
different approach

the tests are KO

the tracks without metadata are imported again and again. Needs fixing.

use url everywhere we refer to tracks or radios

allow to handle tracks without databaseId like the case of a track
without metadata

BUG: 415180