Playlist: adopt simpler and more compact track/disc numbering and playback indicator
Elisa's playlist items currently use a bunch of very complicated code to produce a large amount of whitespace to the left of the track/album numbers that contains only the playback indicator, and aligns the right edge of the album art with the right edge of the numbers. Unfortunately this results in a lot of wasted space in the very narrow space of the playlist, and many song titles are elided.

This MR changes the visual style to reduce the whitespace and replace the track/disc number indicator with the playback indicator for the currently playing song, saving a ton of space. Doing this allows the elimination of a great deal of complex and fragile layouting code.