avoid Elisa prevent session close
This has to do with session management and the fact that when closed
the Elisa window hides and refuse to be closed when system tray icon
is configured active.

https://doc.qt.io/qt-5/qsessionmanager.html is the main input link in
the Qt documentation.

This patch detects that the session is being closed and allow the window
to be closed instead of hidden. That way, Elisa will not signal that the
window refused to close and the session manager will close the full
session.