Fix string puzzle with list and grid view placeholder strings
Using a placeholder for the type of data in the string doesn't work
because of grammatical genders and other differences from English.

The UI code is implemented in a very generic way, which is good
overall, but it prevents us from easily defining a different string
for every that it could be displaying, which would be the best way to
solve the string puzzle.

Accordingly, this commit takes the lazy path by just setting the
placeholder text to a status string with no substitutions. The message
is more generic, but always translatable.