prevent new radio dialog to get data from newly played track
when a create new radio dialog is open and the currently playing track
changes, it will get the data from the newly playing track instead of
being non modified

clean the design of TrackMetadataModel to not have half of the code
needed to build the EditableTrackMetadataModel model. Keep it only on
handling of metadata for tracks.

prepare for proper handling of modifications of metadata of tracks and
not only radios