Add feature to show files in file manager
Now you can show songs' location on disk using your file manager.

This functionality is implemented as new hover buttons in playlist,
list, and grid browser delegates, and a new always-visible button on the Now
Playing footer toolbar.

The button does not appear on grid delegates for files (e.g. in the
Files view) due to a pre-existing bug also affecting the Details button.
This will have to be fixed separately.

BUG: 406471
FIXED-IN: 20.12