Don't look for translated version of "Play-Pause" shortcut
Shortcuts should always use bare string literals instead of translations
because this causes issues in cases where the string has not been
translated in some languages, or when it gets customized before or after
the translation takes place.

This commit fixes the problem by always searching for the string literal.

BUG: 410110
FIXED-IN: 20.12.1