Move remaining track indicator to playlist footer
Right now it can overlap with the text in the HeaderBar. And its
current location there feels kind of random, like it's just floating
in space, unstructured. By moving it to the playlist's footer, we solve
both of those problems and also give it in a more logical home:
alongside the indicator of the total number of tracks. Now the footer
shows two related pieces of information: the total number of tracks,
the the number of remaining tracks.

BUG: 429464
FIXED-IN: 21.04