Fix Radio streams
Adding Radio stream was broken before this. The issue was a check
against the databaseId for new Radiostreams. The check was against a
predefined value for ulong that was not correct anymore. This change
fixes the assumption about the value of not initialized long values by
using hasDatabaseid. Also introduces some tests so this can not
happen again. Hopefully.

Signed-off-by: Sven Marquardt <dev@mail.smarquardt.space>