Prefer m3u8 over m3u as a playlist format
m3u files, when loaded by an audio player, will cause it to choke on
entries with non-ASCII characters--even if the file itself is UTF-8
encoded. To make this work, the filename extension itself must be m3u8,
not m3u.

Elisa already transparently supports files with the .m3u8 extension, but
does not save in this file format by default, and does not list them for
opening in its open dialog.

This commit makes Elisa able to open m3u8 files from its file dialog and
save playlists in that format by default.

CCBUG: 389187