Make track ratings editable inline
This commit makes ratings editable inline from the Now Playing page as
well as the Playlist items. For Playlist items, because there is not a
lot of horizontal space available, the stars only become editable when
you click on a new "Rate this track" button on the existing row of
inline buttons.

BUG: 437099
FIXED-IN: 21.12