Un-hardcode slider appearance
Overriding and hardcoding the background and handle for the duration and
volume sliders resulted in a variety of issues, such as failure to
follow color schemes and fragile custom code. There were various reasons
why this was originally done, such as avoiding a harsh colored appearance
with the former Breeze slider style, and also increasing the size of the
hitbox. The former issue is largely improved by the new "Blue Ocean"
Breeze style in Plasma 5.23, and the latter issue is fixed with a simple
tweak to qqc2-desktop-style; see
https://bugs.kde.org/show_bug.cgi?id=410672.

Accordingly, this commit removes all the custom code and returns to using
a standard slider appearance.

Note that Frameworks 5.86 or current git master is required to avoid a
visual regression. I have noted this in the CMakeLists.txt file and will
formally bump the Frameworks dependency once 5.86 is tagged.

If the new appearance is judged to still be too colorful, we can always
go back to hardcoding just that.