Optional "favorite" style ratings
This commit adds an optional and off-by-default mode to display ratings
as "favorite"/"not favorite" rather than zero to five stars. This
supports people whose rating habits lead them to rate certain songs
with five stars but never use the 1, 2, 3, or 4 star ratings. In this
style, you're basically using the rating system to mark your favorite
songs rather than really rating them. This new mode makes that a bit
more natural for people who prefer that style.

In terms of implementation, when using this mode, songs with 5 stars are
considered to be "favorite" songs, and songs with any other rating are
considered to be "not favorite". Marking a song as a favorite internally
sets it to 5 stars, and it will appear as a 5-star song in other apps.

Note that I have used the existing star-style "rating" and
"rating-unrated" icons for this mode due to the absence of an adequate
pair of filled/unfilled heart icons in the Breeze icon theme. If those
are later added, it could be switched over to use them.

BUG: 430196
FIXED-IN: 21.12