Fix crash when QUrl in Enqueued files is empty
Fixes the issue where the program crashes when you navigate into a folder where
there are no playable music files, and attempt to hit the "Play" or "Add to
Playlist" buttons on top right navigation bar. In the fix we just return in
case the enqueued file QUrl is empty.

BUG: 428218
FIXED-IN: 21.12.2