Port Playlist inlinemessages to PassiveNotification
We were basically replicating the functionality; might as well just use
the real thing. Then we can benefit from future improvements and have
less custom code.

The connections to the model's signal are moved into ElisaMainWindow so
that they are made only once, not once per instantiation of
MediaPlayListView. That caused logic to be executed twice.

BUG: 447607
FIXED-IN: 22.04