Enable unicode normalization on track data before matching
Normalizes the value of mainValue, artistValue and oneArtist with
Qt provided normalization before it is compared to
mFilterExpression's mFilterText and used for search. mFilterText is
also normalized before it's being matched against.

With this fix the provided "Björk" file in the bug report is correctly
found when searching "Björk" in the search bar.

BUG: 391462
FIXED-IN: 21.12.2