Fix metadata view not updating when switching tabs
We need to check whether to call initializeByUrl or initializeByIdAndUrl
based on whether the track has database id, fixes the issue of
the metadata view being blank after switching tabs.

Also onTrackTypeChanged needs the same check so metadata is properly
updated when switching for example from radio to a track without
metadata.

BUG: 436405