Warn that the fast native indexer doesn't follow symlinks
When using the Baloo-powered fast native indexer, symbolic links are
not followed, due to a longstanding design decision in Baloo itself.
This means that if, for example, you have ~/Music confiured as the only
music location, and within that folder is a symbolic link pointing to
~/otherMusicLocation, none of the music there will be visible. The user
will have to manually add ~/otherMusicLocation to the list of music
locations as well.

This is not at all obvious, and it's not going to change in Baloo
anytime soon as it is a design decision. So let's tell the user about
it in the config dialog so they're aware.

BUG: 451294
FIXED-IN: 22.08