Collapse GridBrowserView into DataGridView
GridBrowserView is used only once, in one other component: DataGridView.
Most of its properties are passed right on as aliases. This adds an
unnecessary layer of abstraction that causes fragility due to the
opportunity for the two to drift out of sync or communicate in
imperative rather than declarative ways.

This situation can be simplified by simply moving the content of
GridBrowserView into DataGridView and then deleting it.