Use real SI unit
Fixing the unit to be consistent with the rest of Plasma (e.g. Dolphin).
Using the letter p to mean "per" is specific to the English language, and not standard notation. This is as advised in kformat.h.
I couldn't find a specific format enum for it, but there may be one.