Add setting to disable scan on startup
Add additional checkbox setting to disable autoscan on startup. This is annoying
for some users and with certain music library setups.

When using this setting, the config window informs you of the need to manually
scan every time new music is added. A new "Scan for new music" menu item is
added to facilitate this.