Change "Show Artist" functionality to be hover-based on desktop
There is precisely one circumstance under which the "Show Artist"
button in the navigation toolbar appears and is  useful: when you have
accessed an album from the "Albums" view in the sidebar. In every other
circumstance, it is redundant because you got to the Album view by
navigating there from the artist view first, so you can just click the
back button. As a result, most of the time that this button is
displayed, it is just taking up space, causing artist and album names
to elide earlier, and not being very relevant.

Removing it entirely is overkill though. So instead, let's make the
artist name in the navigation toolbar clickable, and clicking on it will
take the user to the artist view. This way we can remove the button.
Making the UX hover-based makes it less discoverable, but this is an
acceptable price to pay for such an esoteric and mostly-useless piece of
functionality, and it re-uses users' familiarity with other clickable
labels in the headerbar.

This change is only made for the desktop UI; no changes are made for the
mobile UI, since hover-based UIs don't work at all there.