Fix case sensitivity sorting issue in viewsproxymodel
There was an issue where in the backend the entries are sorted
case insensitively, but when sorting the entries in viewsproxymodel 
it was done case sensitive. This caused that albums and other
entries in the sidebar that started with a lower capital letter would be pushed
to the bottom of the list. This discrepancy made it so you could
click an album on the sidebar and the actual album that opened
would be completely different that you clicked.



Before:

![casesensitivity-before-2022-06-30_14.09.06](/uploads/68f5eb71b224b614b1163b7097a175de/casesensitivity-before-2022-06-30_14.09.06.mp4)

After:

![casesensitivity-after-2022-06-30_14.08.04](/uploads/c7c818d9ce63737d19c802312bd87218/casesensitivity-after-2022-06-30_14.08.04.mp4)


(cherry picked from commit 301c7462ef0d516e7b4896854ed6b9f3d82d3ed8)