Set sidebar section Loader height to avoid inactive headings taking space
Add height to the Loader based on the implicit height of the loaded `item`, in the case that the `item` is null (such as when it is not active), we set the height to 0. This fixes the issue of when you have unloaded the "default" section header, and scroll back up, there would be empty space where the header would have been.

BUG: 456214