Use up to four latest album or track covers as artist cover
When retrieving artist data, find up to four latest album or track covers
and display them as the artist's cover in a grid layout. If four covers
can't be found, display only the latest one.

This should improve the UX of the Artists view, which right now is a sea
of identical icons.

BUG: 406475
FIXED-IN: 22.12