Add simple tests for VolumeSlider
VolumeSlider is a Slider that features a MouseArea inside it. This is a combination that I myself have broken
from time to time. Hence, I add three simple test cases for reacting to clicks, for reacting to dragging while
the mouse is pressed and for reacting to the user's wheel input.

The test case for the wheel input requires to expose the 'wheelEffect' property, which is the hack we use
in MouseArea to avoid getting tickmarks. Its value is 5, because that's the value it had without the property.
Exposing this property makes sure that the test case won't fail if, say, the number 5 was exchanged for
the number 6.