Ensure that the current track becomes the first one when shuffling
Previously, when shuffling tracks, the current track was also moved to a
random position. This is very annoying, especially when its position is
close to the end of the playlist.

This fix puts the current track at the top of the queue and only then
shuffles the rest. This type of shuffling is used, for example, by
Spotify.