Remove obsolete workaround for QML combobox width bug
Not only is it not fixing an issue anymore (the bug it was a workaround
for has since been fixed), but it's now causing them! Remove the
workaround and let the combobox do its own sizing.

BUG: 458215
FIXED-IN: 22.08.1
(cherry picked from commit 0fcddb0d5376788f7eb4c2946f8abe2e2926ba4f)