Support saving playlist with relative paths
Adds the support for saving playlists with relative paths in cases where
the music files are located in the same folder as the playlist to be saved.
This setting is "Prefer relative paths" by default as that seems to be
common with most music players.

A new persistent config `alwaysUseAbsolutePlaylistPaths` is created, which
is then checked in `MediaPlayListProxyModel::savePlaylist`. With a false
value, music files in the same folder as the playlist are pathed with just
the filename, otherwise with absolute paths. With a true value an absolute
path is always used.

BUG: 449262
FIXED-IN: 22.12