Fix the loss of rating stars when changing focus
If you select a rated track in the panel of the current playlist, and
then switch the focus to, for example, some track in the central
panel, the rating display on the panel of the current playlist will
disappear.

This is probably happening because Qt thinks the object is still
selected. This may be right for a program, but not for a human.
Removing !isSelected solves this problem.