Change the order of track control buttons
If the user wants to remove a track from favorites, then he move the
mouse to the rating star, but at the last moment it changes to the
Play button or the Remove button (Depending on the panel). After
clicking on these buttons, not what the user wanted happens.

Kirigami.icon was replaced with a button due to the fact that the
alignment of the control buttons and this icon does not coincide
horizontally, which causes movements.