Remove unnecessary code for opacity
Previously, if you changed the focus from an object evaluated in 0-5
mode, the stars disappeared. This was due to the fact that the object
became transparent if its rating <2.

Most likely, this was once done in order to produce alignment.
However, this broke the evaluation system.

This is no longer required, as the interaction with this widget has
changed in previous commits. Therefore, this parameter has been
removed.