Remove extra separator lines in Playlist
We deliberately turn off separator lines in the list item, but then
manually add more outside of it, in the delegate that contains the list
item code.

This was originally done when introducing the new album headers view
that only appears when tracks from the same album are intermixed with
tracks all from different albums, but in retrospect, it's probably too
visually heavy, and we should only use the standard separators that some
from the list items themselves if we want separator lines here.