Make HeaderBar resizeable
The original goal was to implement a way to hide the header bar, so other components
would have more space. I decided the best way to implement this without breaking other
features/UX was to make the header bar shrinkable to the point where only the media
control bar remains visible.