RatingStar: Turn it into a Control to allow keyboard interaction
Unfortunately, FocusRect is a private qqc2-desktop-style thing which we
can't use, so it uses custom graphics for background component to
indicate its focused state. This rectangle doesn't look too bad though,
and is mostly consistent with what "usual" styles would've drawn anyway.

The component does not accept focus when it's not interactive.