Allow configuring the initial Files view path
Right now it defaults to / so that the user can access any external
devices they may have plugged in with music on them. However this is a
fairly cumbersome location to always start in as it exposes system
components to the user.

Let's allow the user to customize the initial Files view location so
that they can make it start in exactly the place where they keep their
files.

BUG: 468523
FIXED-IN: 23.12