Don't open album artist page when opening an album from a different artist's page
Context: There is an album "NCS: Best of 2017" with the album artist tag set to "Various Artists". One of the tracks on this album is by an artist "Anna Yvette".

Steps:
1. Go to artist page for "Anna Yvette"
2. Click on the album "NCS: Best of 2017"
3. Click the back button once

Previous behaviour:
We are now looking at the artist page for "Various Artists". Pressing the back button does not take us to "Anna Yvette", but instead takes us to the main Artists page. I feel that this is unexpected behaviour.

New behaviour:
We are now looking at the artist page for "Anna Yvette".