Use configure_file to make QML files compatible for KF5/KF6
Some properties got renamed and thus we need compatibility logic.
The configure file is ugly, but better than having a KF5/KF6 state in
parallel that can get out of sync over the span of two years.

Because the qrc file paths are reative to the current directory, the
configured file is put in the source directory.
Otherwise, we'd need to have two QRC files or change the paths to be
absolute.