Delete ViewManager::viewIsLoaded()
This was previously used to open two views consecutively. For example: when
opening an album Elisa would first open the Albums list view, and *then*
open the album itself.

Since this behaviour has just been changed (we always open one view, which
is the view the user requested), viewIsLoaded is effectively useless.