Stop Baloo backend from failing when Baloo indexes non-existing folder
Currently the Baloo backend fails to launch when Baloo is set to explicitly
not index a non-existing folder.

This is because we were converting the exclude folder into a canonical path
and then checking if that path was a parent of the music path. Since the
folder did not exist, this resulted in an empty string which we considered
to be a parent of all paths. Hence the code thought that Baloo could not
index that music path.

Now we use QUrl::isParentOf which is more robust than our previous code