Fix sidebar items not sorting automatically
Previously any new embedded sidebar items would appear at the bottom of
the list, unsorted. Now the sidebar is sorted automatically.

E.g. if the sidebar had the genres [Pop, Soul] and the genre "Rock" was
added then previously it would be like [Pop, Soul, Rock] but now it is
[Pop, Rock, Soul].