ContentView: force ReplaceTransition when replacing the stack
Normally Qt will use the Immediate transition if the stack is empty, but
this behaviour is bugged when using OpacityAnimator. Instead of setting
properties to their `to` value, they are set to their `from` value
(these are defined in the Transition). This results in the initial view
item having an opacity of 0. The workaround is to force a
ReplaceTransition.

See https://bugreports.qt.io/browse/QTBUG-117720

CCBUG: 474579