Metadata form: make these properties aliases to avoid destroying bindings
For the current mobile UI we bind the form's `isModifying` to the View's
own property. Then when clicking the Modify button this binding is destroyed,
so the form has `isModifying === true` but the view has `isModifying ===
false`. I don't think this currently leads to any bugs but it is better
to be safe than sorry.