Don't skip to the end when playing the same track twice in a row
If the QMediaPlayer mediaStatus is EndOfMedia and the source is set to the
same value as it was before, the media player instantly jumps to the end.
This means we effectively skip the track when trying to play for a second
time.

Work around for https://bugreports.qt.io/browse/QTBUG-121355

BUG: 480743