Fix crash when flicking mouse over rating star
Previously the app would crash when flicking the mouse over a rating star.

It was something to do with setting the star icon's layer.enabled when entering/leaving
the hover area. Not sure what the root cause was, maybe setting it a lot
of times in a short period of time.

Now we just use one mouse area for the control and calculate the hovered
rating from the location of the mouse with respect to the width of the control.