Fix qt multimedia backend stopping playback after a track has finished
When reaching the end of a track the QMediaPlayer first sets
playbackState to Stopped and then sets mediaStatus to EndOfMedia.

The ManageAudioPlayer class is currently designed to handle changes the
other way round because that's how the old Qt5 backend worked. I.e. set
EndOfMedia first and then set Stopped. This meant the audio player would
first see the Stopped signal and did not know that we should start
playing the next track immediately.

Now we cache the signals from QMediaPlayer and emit them in the order
expected by ManageAudioPlayer.