Fix missing cover art on Android >= 10
The ALBUM_ART column is deprecated as of API level 29 (Android 10 /
Quince Tart). The replacement is to use ContentResolver#loadThumbnail
to get the thumbnail on the Java side and then convert that into a
QImage.

Also the ALBUM_ID column returns a long, not an int. The conversion was
causing overflow issues which meant we were trying to load the cover art
for albums with nonsensical IDs.